requirements.txt
Flask==3.0.3
flask-babel==4.0.0
flask-Cors==4.0.0
psycopg2-binary
python-dotenv==1.0.0
hashlib